pdf/requirements.txt
python-pptx==0.6.21
tqdm==4.65.0
requests==2.31.0 